add component card
