+++ nextjs/marisolsalazarv-6177/pages/index.jsx
@@ -1,22 +1,7 @@
-import Dog from "@/components/Dog/Dog";
+import Dog from "@/components/Dog";
 import styles from "@/styles/Home.module.css";
-import React, { useEffect, useState } from "react";
 
 export default function Home() {
-  const [dogs, setDogs] = useState([]);
-
-  useEffect(() => {
-    const fetchData = async (res, req) => {
-      try {
-        const dogs = await fetch("/api/dogs").then((res) => res.json());
-        setDogs(dogs);
-      } catch (error) {
-        return res.status(500).json({ message: error.message });
-      }
-    };
-    fetchData();
-  }, []);
-
   return (
     <>
       <title>Dog Breeds</title>
@@ -26,9 +11,7 @@ export default function Home() {
         DOG BREEDS
       </h1>
       <main className={`${styles.body}`}>
-        {dogs.map((dog, index) => {
-          return <Dog key={index} title={dog.title} img={dog.img} />;
-        })}
+        <Dog />
       </main>
     </>
   );
